requirements.txt
pika
python-dotenv
azure-servicebus
